Move search_by_otu_id() to .catalog.catalog
